Re-organize the base layouts & Merged Toolbars
Large re-organization and reworking of flows. The "main" panel is now a lot less coupled
to the rest of the components and it will be much easier to add and remove panels.

Additionally merged the multiple toolbars and pushed them into the header to gain some
vertical space that is much needed.

Incidentally added a "Back" button for flashing and replace the "Flash" button when in
configurator mode to be "download"