Convert Project to Typescipt
This commit moves the entire project (with the exception of one file which has
typings and will be replaced by redux) to typescript. With this commit there is
a good first pass at typing the entire project. It is definitely not perfect
and there are more than a few forced conversions taking place. Over time these
will be replaced with proper type guards and/or algebraic types to ensure the
need for direct coercion is not necessary.

As to why this change was made, this is in prep to remove the current global
state solution that is currently in and replace it with redux. The current
solution has already been stretched thin and in order to make such a sweeping
change without introducing massive breaks proper typing is necessary.
Additionally, changes to underlying objects to support new functionality (such
as non-RGB lighting and Ergodox Flashing) is currently too slow and haphazard.
Again a solution typing will solve.